refactor: database schema separations
